+++ ante/gov_ante_test.go
@@ -11,11 +11,11 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 
 	"github.com/atomone-hub/govgen/ante"
 	govgenapp "github.com/atomone-hub/govgen/app"
 	govgenhelpers "github.com/atomone-hub/govgen/app/helpers"
+	govtypes "github.com/atomone-hub/govgen/x/gov/types"
 )
 
 var (