+++ x/gov/handler_test.go
@@ -1,0 +1,25 @@
+package gov_test
+
+import (
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+
+	"github.com/cosmos/cosmos-sdk/testutil/testdata"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/atomone-hub/govgen/x/gov"
+	"github.com/atomone-hub/govgen/x/gov/keeper"
+)
+
+func TestInvalidMsg(t *testing.T) {
+	k := keeper.Keeper{}
+	h := gov.NewHandler(k)
+
+	res, err := h(sdk.NewContext(nil, tmproto.Header{}, false, nil), testdata.NewTestMsg())
+	require.Error(t, err)
+	require.Nil(t, res)
+	require.True(t, strings.Contains(err.Error(), "unrecognized gov message type"))
+}