+++ x/gov/module_test.go
@@ -1,0 +1,29 @@
+package gov_test
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	abcitypes "github.com/tendermint/tendermint/abci/types"
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+
+	govgenhelpers "github.com/atomone-hub/govgen/app/helpers"
+	"github.com/atomone-hub/govgen/x/gov/types"
+)
+
+func TestItCreatesModuleAccountOnInitBlock(t *testing.T) {
+	app := govgenhelpers.SetupNoValset(false)
+	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
+
+	app.InitChain(
+		abcitypes.RequestInitChain{
+			AppStateBytes: []byte("{}"),
+			ChainId:       "test-chain-id",
+		},
+	)
+
+	acc := app.AccountKeeper.GetAccount(ctx, authtypes.NewModuleAddress(types.ModuleName))
+	require.NotNil(t, acc)
+}