+++ CHANGELOG.md
@@ -12,5 +12,13 @@
 
 ### FEATURES
 
-### STATE BREAKING
+* Forked `x/gov` module from Cosmos SDK v0.46.16 for custom modifications
+  ([#4](https://github.com/atomone-hub/govgen/pull/4)).
+* Removed ability for validators to vote on proposals with delegations, they can only use their own stake
+  ([#20](https://github.com/atomone-hub/govgen/pull/20)).
+* Remove community spend proposal
+  ([#13](https://github.com/atomone-hub/govgen/pull/13)).
+* Adapt voting period to proposal type.
+  ([#16](https://github.com/atomone-hub/govgen/pull/16)).
 
+### STATE BREAKING